Rename component to <Mentoring.Discussion />
