+++ app/controllers/test/components/mentoring/discussion_controller.rb
@@ -1,4 +1,4 @@
-class Test::Components::Mentoring::MentorDiscussionController < Test::BaseController
+class Test::Components::Mentoring::DiscussionController < Test::BaseController
   def show
     @discussion = Solution::MentorDiscussion.find(params[:discussion_id])
   end
