+++ app/javascript/components/mentoring/discussion/Guidance.tsx
