+++ app/javascript/components/mentoring/discussion/Scratchpad.tsx
