+++ app/views/test/components/mentoring/discussion/show.html.haml
@@ -1,0 +1,1 @@
+= ReactComponents::Mentoring::Discussion.new(@discussion)
