--- app/views/test/components/mentoring/mentor_discussion/show.html.haml
@@ -1,1 +1,0 @@
-= ReactComponents::Mentoring::MentorDiscussion.new(@discussion)
