+++ test/javascript/components/mentoring/discussion/DiscussionPostList.test.js
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react'
 import { rest } from 'msw'
 import { setupServer } from 'msw/node'
 import '@testing-library/jest-dom/extend-expect'
-import { DiscussionPostList } from '../../../../../app/javascript/components/mentoring/mentor-discussion/DiscussionPostList'
+import { DiscussionPostList } from '../../../../../app/javascript/components/mentoring/discussion/DiscussionPostList'
 
 test('displays all posts', async () => {
   const twoDaysAgo = new Date(new Date() - 1000 * 60 * 60 * 24 * 2)
